refactor: move webhook to another controller
